Fixed utils test
